Extract unrelated date conversion functionality from date block
[MAILPOET-2665]
