+++ tests/unit/Form/Util/DateConverterTest.php
@@ -2,16 +2,16 @@
 
 namespace MailPoet\Test\Form\Block;
 
-use MailPoet\Form\Block\Date;
+use MailPoet\Form\Util\DateConverter;
 
-class DateTest extends \MailPoetUnitTest {
+class DateConverterTest extends \MailPoetUnitTest {
 
-  /** @var Date */
-  private $date;
+  /** @var DateConverter */
+  private $dateConverter;
 
   public function _before() {
     parent::_before();
-    $this->date = new Date();
+    $this->dateConverter = new DateConverter();
   }
 
   public function testItCanConvertDateMonthYearFormatToDatetime() {
@@ -22,7 +22,7 @@ class DateTest extends \MailPoetUnitTest {
       'YYYY/DD/MM' => '2016/10/05',
     ];
     foreach ($date as $dateFormat => $date) {
-      expect($this->date->convertDateToDatetime($date, $dateFormat))
+      expect($this->dateConverter->convertDateToDatetime($date, $dateFormat))
         ->equals('2016-05-10 00:00:00');
     }
   }
@@ -33,37 +33,37 @@ class DateTest extends \MailPoetUnitTest {
       'YYYY/MM' => '2016/05',
     ];
     foreach ($date as $dateFormat => $date) {
-      expect($this->date->convertDATEToDatetime($date, $dateFormat))
+      expect($this->dateConverter->convertDateToDatetime($date, $dateFormat))
         ->equals('2016-05-01 00:00:00');
     }
   }
 
   public function testItCanConvertMonthToDatetime() {
     $currentYear = date('Y');
-    expect($this->date->convertDateToDatetime('05', 'MM'))
+    expect($this->dateConverter->convertDateToDatetime('05', 'MM'))
       ->equals(sprintf('%s-05-01 00:00:00', $currentYear));
   }
 
   public function testItCanConvertYearToDatetime() {
-    expect($this->date->convertDateToDatetime('2016', 'YYYY'))
+    expect($this->dateConverter->convertDateToDatetime('2016', 'YYYY'))
       ->equals('2016-01-01 00:00:00');
   }
 
   public function testItCanConvertDatetimeToDatetime() {
-    expect($this->date->convertDateToDatetime('2016-05-10 00:00:00', 'datetime'))
+    expect($this->dateConverter->convertDateToDatetime('2016-05-10 00:00:00', 'datetime'))
       ->equals('2016-05-10 00:00:00');
   }
 
   public function testItCanClearDate() {
-    expect($this->date->convertDateToDatetime('0/10/5', 'YYYY/MM/DD'))
+    expect($this->dateConverter->convertDateToDatetime('0/10/5', 'YYYY/MM/DD'))
       ->equals(date('Y') . '-10-05 00:00:00');
-    expect($this->date->convertDateToDatetime('0/0/5', 'YYYY/MM/DD'))
+    expect($this->dateConverter->convertDateToDatetime('0/0/5', 'YYYY/MM/DD'))
       ->equals(date('Y') . '-' . date('m') . '-05 00:00:00');
-    expect($this->date->convertDateToDatetime('0/0/0', 'YYYY/MM/DD'))
+    expect($this->dateConverter->convertDateToDatetime('0/0/0', 'YYYY/MM/DD'))
       ->equals('');
-    expect($this->date->convertDateToDatetime('0', 'YYYY'))
+    expect($this->dateConverter->convertDateToDatetime('0', 'YYYY'))
       ->equals('');
-    expect($this->date->convertDateToDatetime('0', 'MM'))
+    expect($this->dateConverter->convertDateToDatetime('0', 'MM'))
       ->equals('');
   }
 }
